+++ src/components/Sidebar/Sidebar.css
@@ -1,10 +1,5 @@
-.sidebar {
-  min-width: 270px;
-  /* padding-top: 30px; */
-  /* padding: 70px 70px 0 70px; */
-}
-
 .sidebar__list {
+  min-width: 270px;
   max-width: 310px;
   list-style: none;
   background-color: #212526;
