+++ src/components/Sidebar/Sidebar.tsx
@@ -8,14 +8,12 @@ const Sidebar = memo(function SideBar() {
   const myCategories = useAppSelector(state => state.categories);
 
   return (
-    <section className="sidebar">
-      <ul className="sidebar__list">
-        <li className="sidebar__item">Все товары</li>
-        {myCategories.map((category: Category) => (
-          <SidebarItem key={category.id} title={category?.name} />
-        ))}
-      </ul>
-    </section>
+    <ul className="sidebar__list">
+      <li className="sidebar__item">Все товары</li>
+      {myCategories.map((category: Category) => (
+        <SidebarItem key={category.id} title={category?.name} />
+      ))}
+    </ul>
   );
 });
