+++ src/pages/Admin/Products/CreateProductForm/CreateProductForm.css
@@ -80,7 +80,7 @@
   border: none;
   border-radius: 5px;
   font-family: Raleway;
-  resize: vertical;
+  resize: none;
 }
 
 .product-form__textarea:focus {
@@ -123,7 +123,7 @@
 
 .product-form__file-label:hover {
   opacity: 0.8;
-  background-color: #f05a00;
+  background: #f05a00;
 }
 
 .product-form__file-label:hover .product-form__file-button-text {
