+++ src/pages/Admin/Products/CreateProductForm/CreateProductForm.tsx
@@ -1,11 +1,12 @@
 import './CreateProductForm.css';
-import { useState, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { useCreateProductMutation } from '../../../../Api/productApi';
 import { ProductFormData } from '../../../../types/productType';
 import usePosterFileInput from '../../../../hooks/usePosterFileInput';
 import { useAppSelector, useAppDispatch } from '../../../../hooks/redux';
 import { setError, clearError } from '../../../../redux/errorSlice';
+import AddDetailsProduct from '../AddDetailsProduct/AddDetailsProduct';
 
 type CreateProductFormProps = {
   submitBtnName?: string;
@@ -24,15 +25,6 @@ function CreateProductForm({ submitBtnName, handleCloseModal }: CreateProductFor
   const dispatch = useAppDispatch();
   const errorApi = useAppSelector(state => state.error.message);
   const myCategories = useAppSelector(state => state.categories);
-  const [detail, setDetail] = useState([]);
-
-  const addDetails = () => {
-    setDetail([...detail, { title: '', description: '', number: Date.now() }]);
-  };
-
-  const deleteDetails = number => {
-    setDetail(detail.filter(i => i.number !== number));
-  };
 
   useEffect(() => {
     dispatch(clearError());
@@ -147,60 +139,7 @@ function CreateProductForm({ submitBtnName, handleCloseModal }: CreateProductFor
         ></textarea>
         <span className="product-form__input_error">{errors?.description?.message}</span>
       </label>
-      <button className="product-form__add-details" type="button" onClick={addDetails}>
-        Добавить характеристику
-      </button>
-      {detail.map(i => (
-        <div className="product-form__box" key={i.number}>
-          <input
-            className="product-form__input product-form__price"
-            id="detail"
-            placeholder="Характеристика"
-            type="text"
-            {...register('name', {
-              required: 'Поле обязательно для заполнения',
-              minLength: {
-                value: 2,
-                message: 'Введите не менее 2 символов',
-              },
-              maxLength: {
-                value: 30,
-                message: 'Введите менее 30 символов',
-              },
-              pattern: {
-                value: /^[a-zA-Z-0-9\u0430-\u044f\u0410-\u042fёЁ\s]*$/,
-                message: 'Введите корректное имя',
-              },
-            })}
-          />
-          <span className="product-form__input_error">{errors?.price?.message}</span>
-          <input
-            className="product-form__input product-form__price"
-            id="detailDescr"
-            placeholder="Описание"
-            type="text"
-            {...register('name', {
-              required: 'Поле обязательно для заполнения',
-              minLength: {
-                value: 2,
-                message: 'Введите не менее 2 символов',
-              },
-              maxLength: {
-                value: 30,
-                message: 'Введите менее 30 символов',
-              },
-              pattern: {
-                value: /^[a-zA-Z-0-9\u0430-\u044f\u0410-\u042fёЁ\s]*$/,
-                message: 'Введите корректное имя',
-              },
-            })}
-          />
-          <span className="product-form__input_error">{errors?.price?.message}</span>
-          <button className="product-form__add-details" type="button" onClick={() => deleteDetails(i.number)}>
-            Удалить
-          </button>
-        </div>
-      ))}
+      <AddDetailsProduct />
       <input
         className="product-form__file"
         type="file"
